# specs/001-user-authentication-onboarding/contracts/api-spec.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "LeaveLab Authentication API",
    "description": "Authentication and user management API for LeaveLab platform",
    "version": "1.0.0",
    "contact": {
      "name": "LeaveLab API Support",
      "email": "api@leavelab.com"
    }
  },
  "servers": [
    {
      "url": "https://leavelab.com/api/v1",
      "description": "Production server"
    },
    {
      "url": "http://localhost:3000/api/v1",
      "description": "Development server"
    }
  ],
  "paths": {
...
    "/auth/logout": {
      "post": {
        "summary": "Logout current user",
        "description": "Invalidate current session and log out user",
        "operationId": "logout",
        "tags": ["Authentication"],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Logout successful",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string",
                      "example": "Logged out successfully"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Not authenticated",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/auth/update-password": {
      "post": {
        "summary": "Update password with reset token",
        "description": "Set new password using reset token from email",
        "operationId": "updatePassword",
        "tags": ["Authentication"],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdatePasswordRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Password updated successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string",
                      "example": "Password updated successfully"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Invalid or expired token",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/profile/avatar": {
      "post": {
        "summary": "Upload profile avatar",
        "description": "Upload or update user's profile picture. Max 5MB, JPG/PNG/WebP only.",
        "operationId": "uploadAvatar",
        "tags": ["Profile"],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "properties": {
                  "file": {
                    "type": "string",
                    "format": "binary",
                    "description": "Image file (JPG, PNG, or WebP, max 5MB)"
                  }
                },
                "required": ["file"]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Avatar uploaded successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "avatar_url": {
                      "type": "string",
                      "format": "uri",
                      "example": "https://storage.leavelab.com/avatars/user-123-1696723200000.webp"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Invalid file (wrong format, too large, etc.)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "Not authenticated",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/account/change-password": {
      "post": {
        "summary": "Change password (authenticated)",
        "description": "Change password for authenticated user. Requires current password.",
        "operationId": "changePassword",
        "tags": ["Account"],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ChangePasswordRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Password changed successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string",
                      "example": "Password changed successfully"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Invalid request (weak password, wrong current password, etc.)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "Not authenticated",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/account/export": {
      "get": {
        "summary": "Export user data (GDPR)",
        "description": "Download all user data in JSON format",
        "operationId": "exportData",
        "tags": ["Account"],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Data export successful",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserDataExport"
                }
              }
            }
          },
          "401": {
            "description": "Not authenticated",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "schemas": {
...
      "UpdatePasswordRequest": {
        "type": "object",
        "required": ["token", "password"],
        "properties": {
          "token": {
            "type": "string",
            "description": "Password reset token from email"
          },
          "password": {
            "type": "string",
            "format": "password",
            "minLength": 8,
            "example": "NewSecurePass123!"
          }
        }
      },
...
      "ChangePasswordRequest": {
        "type": "object",
        "required": ["current_password", "new_password"],
        "properties": {
          "current_password": {
            "type": "string",
            "format": "password"
          },
          "new_password": {
            "type": "string",
            "format": "password",
            "minLength": 8
          }
        }
      },
      "User": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "example": "a1b2c3d4-e5f6-7890-abcd-ef1234567890"
          },
          "email": {
            "type": "string",
            "format": "email",
            "example": "user@example.com"
          },
          "email_confirmed": {
            "type": "boolean",
            "example": true
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "example": "2025-10-08T12:00:00Z"
          },
          "last_sign_in_at": {
            "type": "string",
            "format": "date-time",
            "example": "2025-10-08T14:30:00Z"
          }
        }
      },
...
      "UserDataExport": {
        "type": "object",
        "properties": {
          "user": {
            "$ref": "#/components/schemas/User"
          },
          "profile": {
            "type": "object",
            "description": "User profile data"
          },
          "auth_logs": {
            "type": "array",
            "description": "Authentication history",
            "items": {
              "type": "object"
            }
          }
        }
      },
      "ErrorResponse": {
        "type": "object",
        "properties": {
          "error": {
            "type": "string",
            "example": "Invalid email or password"
          },
          "code": {
            "type": "string",
            "example": "AUTH_INVALID_CREDENTIALS"
          },
          "details": {
            "type": "object",
            "additionalProperties": true,
            "description": "Additional error context"
          }
        },
        "required": ["error"]
      }
    }
  },
...
}